0.0.1
- Floating numbers support
- Division by zero handling
- REPL improvements

+++ eval/include/eval.h
@@ -2,4 +2,4 @@
 
 #include "parser/include/ast.h"
 
-int eval(ASTNode *node);
+float eval(ASTNode *node);
